feat(sdk): `aws.QueueRef` (#6001)
Introduce `aws.QueueRef` as a way to add an external queue reference to an application:

```js
bring aws;
bring cloud;

let myQueue = new aws.QueueRef("arn:aws:sqs:...."); // <-- ARN of an existing SQS queue

new cloud.Function(inflight () => {
  myQueue.push("new message");
});
```

This works both when running in the simulator and when running on AWS.

The Wing Console also includes some metadata about the queue such as the queue URL and the AWS Console URL:

![image](https://github.com/winglang/wing/assets/598796/fec3a609-6f7e-42c9-93ec-1768c197ca07)

### Issues

* Related to #5879 (only implements queue and we need additional resources).
* Related to #5170


### Misc

Added an option `link: bool` to `ui.FIeld` to indicate that the value returned is a link. We should make this a clickable link in the Wing Console.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.