fix(sdk)!: `util.waitUntil` is silent if condition is not met (#6008)
Closes #6006

`waitUntil()` is a built-in function in the `util` module for automatically polling and waiting until a condition is true (or until a timeout has been reached). From looking at how it was used in winglibs and example programs within the Winglang test suite, we saw there were a lot of cases where `waitUntil()` was called and the return value was silently ignored.

To make the API more intuitive, this PR changes the default behavior so that the function throws if the condition is never met within the timeout period, so it's no longer necessary to wrap the function in an `assert()` call for the most common use case. For the less-common case where you want to inspect whether the condition was reached and perform some action, it's now possible to set the `throws` option to `false` -- which is equivalent to switching to the old behavior.

BREAKING CHANGE: `util.waitUntil()` now throws if the condition is not met within the timeout period. To turn off this behavior, you can pass `throws: false` as a named option to the function.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.