fix: language server occasionally duplicates types a fails compilation (#6009)
Fixes #6004, I was able to replicate consistently and see that this change fixes it

Instead of attempting to reuse `Types`, let's just create a new one. This should cause all the existing types and symbol environments to be dropped, but that's okay because we immediately populate the new one with data.

No clue how to add a real test for this. The current rust-based tests don't have a multi-file system and I'm not sure this change is worth that effort to improve (as nice as it would be)

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.