chore: deprecate `LegacySandbox`
Migrate `cloud.Service` to use `Sandbox` instead of `LegacySandbox`.

Since we can't transmit a JavaScript function across the process boundary, I've modified the service entrypoint to expose `start()` and `stop()` methods which are called by `cloud.Service` in response to the respective lifecycle events.

For now, I've kept the contract of `cloud.Service` in tact, but I think we will want to change it so it accepts an object that has an `inflight start()` and and `inflight stop()` methods, but this is left for a future PR.