chore: deprecate `LegacySandbox` (#6048)
Migrate `cloud.Service` to use `Sandbox` instead of `LegacySandbox`.

Since we can't transmit a JavaScript function across the process boundary, I've modified the service entrypoint to expose `start()` and `stop()` methods which are called by `cloud.Service` in response to the respective lifecycle events.

For now, I've kept the contract of `cloud.Service` in tact, but I think we will want to change it so it accepts an object that has an `inflight start()` and and `inflight stop()` methods, but this is left for a future PR.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.