feat(cli): cloud test snapshots
When running `wing test -t TARGET` for a cloud target (e.g. `tf-aws`, etc), the framework will capture a
snapshot of the synthesized output and store it next to the test file (only if the tests passed on the cloud).

Then, when running the same command within a CI system (`CI=1`), the framework will not actually deploy,
but will only validate that the output hasn't changed.

A few CLI switches were added to control this behavior:

* `--snapshots=auto` - auto-detect based on CI flag
* `--snapshots=never` - disables snapshots
* `--snapshots=dry` - will skip deployment and only update the snapshots.
* `--snapshots=wet` - forces a deployment even if `CI=1`.
* `--snapshots=assert` - only asserts that the snapshots have not changed

Updated the `tf-aws-test.yml` workflow to use `--snapshots=never` since we actually
want to run these tests on the cloud.