fix(sdk): cloud.Service shutdown code is ignored when Wing Console is stopped (#6089)
Fixes #6018

If the Wing Console is terminated by Ctrl + C, it gracefully waits for resources to be shut down (thanks to https://github.com/winglang/wing/pull/6051). However there's a latent bug since the kill signal is still propagated to child processes, like those running inflight code for cloud.Service. This changes the behavior so the kill signal is longer propagated, and the simulator will be fully responsible for shutting down its child processes.

If this doesn't solve all our issues (or runaway processes occur in some other ways), then we might need some kind of daemon for tracking processes created by the Wing simulator - but I'm hoping we might be able to avoid this kind of complexity for now.

## Checklist

- [ ] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [ ] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.