fix(sdk): lift/permissions errors are not caught in simulator (#6079)
Fixes #6002

To support the needs of cloud applications with many resources, Wing's compiler performs static analysis to identify all of the infrastructure resources that are used by different bits of runtime code. (In Wing parlance these are usually called preflight objects, and inflight functions, respectively). By identifying these relationships, the compiler can automatically generate applications whose resources have least-privileged access. For example, database read/write permissions only need to be added to the compute resources that perform queries on it.

This analysis is sufficient for many programs. But in general, obtaining the minimal set of resources needed for arbitrary inflight code is, in the worst case, as hard as the [halting problem](https://en.wikipedia.org/wiki/Halting_problem). To illustrate, the inflight code below only needs access to the preflight object `buckets[0]`, but it's impossible for any reasonable compiler to deduce this.

```js
bring cloud;

let buckets = [
  new cloud.Bucket() as "b1",
  new cloud.Bucket() as "b2"
];

new cloud.Function(inflight () => {
  let var i = 10;
  while i > 0 {
    i -= 1;
  }
  buckets.at(i).put("key", "value");
});
```

Even if Wing's analyzer did provide a [covering set](https://en.wikipedia.org/wiki/Set_cover_problem) of the resources and permissions needed by inflight code (in other words, a set of permissions that might be broader than necessary), there are cases when users will want or need to scope down permissions.

https://github.com/winglang/wing/pull/5935 addresses some of these needs by letting the user annotate functions with additional lifting information:

```js
new cloud.Function(inflight () => {
  lift(buckets.at(1), ["put"]);
  let var i = 10;
  while i > 0 {
    i -= 1;
  }
  buckets.at(i).put("key", "value");
});
```

This was an important change to unblock many use cases. Unfortunately, a side effect is that we've lost the guarantee that every program that compiles and works in the Wing simulator will also work in the cloud. That's because the simulator doesn't simulate permissions -- so even if bad `lift()` annotations are provided by the user, a program will still work in the simulator. Ouch!

As a rule of thumb, we want to maximize the extent to which Wing programs working on the simulator result in the same programs working on the cloud. Even though we can't catch these errors at compile time, they can be detected at runtime. To that end, this PR implements a basic permissions model within the Wing simulator. Whenever a resource tries calling a method on another resource, the API call will first be checked against a list of permissions (generated during preflight) to see if the request should go through. If not, an error will be raised to the caller. The end result is that as long as you're able to test the code path with a unit test, it should still work locally.

Here's an example of the error that will be surfaced to the user during `wing test` or in the Wing Console:

```
Error: Resource "root/Default/Function" does not have permission to perform operation "put" on resource "root/Default/b1".
   --> bucket.main.w:14:3
   | while i > 0 {
   |   i -= 1;
   | }
14 | buckets.at(i).put("key", "value");
   | ^
at /Users/chrisr/dev/wing-test/bucket.main.w:14:3
```

-----

### Implementation notes

**The good:** The main implementation is built around a new resource named `sim.Policy`. It stores a "principal" (a resource the policy applies to) and a list of statements (a list of resources and method names). It's probably not the only way to model this, but it felt like a decent starting point given that permissions are also represented with resources within CloudFormation and Terraform on most major cloud providers.

Policy checking is also straightforward. All requests between resources currently go through the simulator's internal HTTP server. We've added an additional piece of information that identifies who is the resource that's performing the method call, which is sent to the server.

**The mid:** How to store the information from `sim.Policy` in the simulator is an open question. This PR models `sim.Policy` like a resource (why not?). Except, nobody communicates with the policy directly. Instead, when the simulator is initializing or shutting down resources, if it sees a resource is a policy, it saves the resolved policy information into a policy registry. When we're checking if an API call should be allowed or not, it performs a lookup in the policy registry.

But does that mean `sim.Policy` has no effect on other resources? Not quite. First, an aside...

> Within the simulator, a resource can be identified in two ways. The first is with their construct path (a string like `root/MyApp/Storage/Bucket`), and the second is with a simulation-unique ID aka "handle" (a string like `sim-42`), which serves as an analog to physical resource names on cloud providers.

> To make this concrete, imagine you have a Wing app with a `sim.Container`, and you modify the container's image. When the simulator is updating to the new version of your app, the construct path for the resource will stay the same, but the resource itself needs to be destroyed and recreated, which results in a fresh simulation ID.

> Today this difference between these IDs isn't strictly necessary - it's possible we could use construct paths everywhere in the simulator. But using these IDs helps us detect dependency relationships in the resource graph, and if we wanted to model other kinds of application updates in the simulator (for example, Terraform's [`create_before_destroy` property](https://developer.hashicorp.com/terraform/language/meta-arguments/lifecycle#create_before_destroy)), then this distinction suddenly becomes very important. So for now I'm trying to keep it.

Each `sim.Policy` refers to resources via their simulation handles, as we do elsewhere in the simulator. But this means a `sim.Policy` resource can't be created until all of the resources it refers to have been created. This introduces some interesting edge cases (for example, how does a `cloud.Service` call other resources during start-up when its `sim.Policy` can't be created until `cloud.Service` has finished starting up??) I managed to solve most of these cases with auxiliary resources that run after the initial resource's creation.

**The bad:** In order for resources to make requests to the simulator server with their own handles, they need to know their own handles. Unfortunately, to expose this information in `ISimulatorContext`, I needed to refactor inflight simulation classes so that simulator context is passed to `init()` instead of to the JavaScript constructor. This results in some uglier code, but I don't think the distinction between constructors and `init()` will be relevant once more sim resources are implemented in Wing.

## Checklist

- [ ] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [ ] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.