fix: missing uncaught inflight exceptions when debugging (#6152)
During normal usage, when an unhandled exception occurs inflight we will prepare a stack trace for the user to show them where things went wrong. When debugging, the `throw` itself tells you it happened first. Since we were technically catching all inflight errors, the useful concept of "break on uncaught exception" didn't work.

With this change, the original errors remain uncaught and we can still communicate that error with the parent process.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.