chore(sdk): refactor inflight hosts to use inline policies in sim (#6123)
Per a suggestion from #6079, this PR refactors the inflight host resources in the SDK (cloud.Service and cloud.Function) so that their policies are inlined. Other resources like `cloud.Api`, `cloud.Bucket`, `cloud.Topic`, `cloud.Schedule`, and `cloud.Queue` continue to use `sim.Policy`.

In general, `sim.Policy` is more flexible in the current implementation since using it doesn't automatically introduce a dependency on the resource(s) permissions are granted to. For example, a `Topic` can have external policy specifying that it can invoke a `Function`, without necessarily introducing a dependency between the `Topic` and `Function`.

But for `Function` or `Service`, this concern isn't as relevant since all of the permissions they use are derived from the inflight function(s) passed to them. So if a `Function` needs permissions for `Bucket`, it's going to also need an environment variable providing the name of the bucket, so a dependency between the two resources will be required either way.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.