feat(cli): cloud test snapshots (#6163)
When running `wing test -t TARGET` for a cloud target (e.g. `tf-aws`, etc), the framework will capture a snapshot of the synthesized output and store it next to the test file (only if the tests passed on the cloud).

Then, when running the same command within a CI system (`CI=1`), the framework will not actually deploy, but will only validate that the output hasn't changed.

A few CLI switches were added to control this behavior:

* `--snapshots=auto`   - auto-detect based on CI flag
* `--snapshots=never`  - disables snapshots
* `--snapshots=deploy` - forces a deployment even if `CI=1`.
* `--snapshots=update` - will skip deployment and only update the snapshots.
* `--snapshots=assert` - only asserts that the snapshots have not changed

Updated the `tf-aws-test.yml` workflow to use `--snapshots=deploy` since we actually
want to run these tests on the cloud.

Results look like this:

![Screenshot 2024-04-09 at 16 04 05](https://github.com/winglang/wing/assets/598796/58505312-2091-420e-aefa-52904d69bbad)




## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.