feat(sdk): stream logs in `wing test` (#6205)
This PR introduces log streaming for the `wing test` command is used with the simulator. By default, any `log` statements printed by the user will be automatically streamed while the tests are running. (The first line of each log includes a string identifying which test it's associated with). More granular trace events from the individual resources can be displayed with the `--debug` flag.

The order of logs can vary and they each include timestamps, so the `--no-stream` flag can be passed to fall back to the old output style. This is used within the monorepo in places where we're snapshotting CLI output for catching regressions.

https://github.com/winglang/wing/assets/5008987/02c9e769-f6b6-4b17-8e5f-2418fcee7531

Closes #2096

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.