fix: invokeAsync can fail during simulator cleanup (#6254)
hangar sometimes run into this:

`InvokeAsync (payload=undefined) failure. Error: Process exited with code null, signal SIGTERM`

I believe this can happen if InvokeAsync is running while it receive the SIGTERM cleanup. This should not be considered a failure, just a normal part of the cleanup process.

The test relies on a sleep so I'm open to better ideas, but it does replicate the issue on `main`

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.