chore: improve log filtering (#6262)
When `wing test` is run against the simulator, logs are streamed to the CLI output. But for technical reasons not all logs are associated with tests. For example, there's a singleton class named `TestRunner` which can have an overridden implementation per target. This is a side effect of how we currently implement test isolation as construct subtrees.

As a small improvement, when events are emitted from these resources, this PR changes the current behavior so they're logged to standard output with the "(no test)" label instead of being ignored.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.