fix(sdk): store simulator paths as absolute path (#6299)
Normalize the code within the SDK so that the paths to various simulator files are stored as absolute paths.

This is causing an issue where the feature added in https://github.com/winglang/wing/pull/6295 only works when running `wing it` and not in `wing test` because `wing test` initializes the simulator with a relative path.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.