fix: cli cleanup is skipped when multiple signals are sent (#6374)
Finally was able to reproduce the issue locally. Depending on how it is executed, it is possible the CLI will receive multiple SIGINT signals when a user hits CTRL^C (even when pressed once).

Using process.once handled the first signal and then started our cleanup process. However, the second signal came along and no longer had a handler because the first process.once was used up. So then node returns to its default behavior, which is to immediately process.exit(). process.exit() Stops the process kills the process without waiting for the event loop to finish so thus the cleanup process is interrupted.

Other changes:
- Removed `exit` handling because that does not allow us to schedule any additional async work, so it's pointless
- Added SIGHUP handling

Reproducing: I noticed this happening with a globally installed version of the CLI, but no with my local dev version. So to test this fix, I manually replaced the compiled .js for this particular util file in the global version. I do not know why this only happens with the global version.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.