fix(sdk): container logs are not emitted (and more)
Change the way we are running containers to capture logs from docker commands and from running containers.

Make some improvements to how container lifecycle is managed by running it without `--detach` and monitoring the `docker` child process.

Add some initial support for log levels, because, come on, currently errors emitted from the simulator or resources are hidden by default because they are treated as "verbose". We need to follow up with a console change for this, but this sets the initial groundwork.