feat(sdk): containers logs and improved simulator stability (#6408)
## Container improvements

Revisit how we are running containers in `sim.Container` to improve robustness and capture logs from docker build/pull commands and from running containers.

To do that we've changed how container lifecycle is managed by running it without `--detach` and monitoring the child process for exit codes, errors and stdio capture.

## Simulator stability improvements

Additionally, this fixes a few long standing issues in the simulator which caused it to hang when one of the resources failed to start (e.g. due to unresolvable token). To fix this, we track failures and if there were any failing resources, we stop all the started resources and throw an error. Additionally, in the CLI we needed to catch this error, stop the spinner and bail out from the test execution routine. I believe this fixes #6052 and #5426.

## Logging improvements

Add some initial support for log levels, because, come on, currently errors emitted from the simulator or resources are hidden by default because they are treated as "verbose". This includes the relevant Console fixes.

Tweak a little how logs are displayed when `DEBUG=1` so that log entries are separated by a newline and removed the indentation of the first line (basically like jest/vitest is doing).

## Other improvements

Hide the `cloud.Endpoint` nodes from the graph because it was a mistake (mine) to suggest to show them.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.