chore: fix e2e test in wrong folder (#6436)
`pnpm turbo test --filter=wingc` is failing on fresh builds. The `invalid` dir is for tests that do not compile, but the test added in #6408 compiles fine, so I moved it to the `error` dir instead (maybe these are in need of renaming?). I still haven't figure out why the test failure wasn't caught in CI/CD.

## Checklist

- [ ] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [ ] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.