feat(vscode): rename refactor (#6461)
## Checklist
fixes: [#4596](https://github.com/winglang/wing/issues/4596)
https://www.loom.com/share/234a65f04ecb471b84621aa0cf27f66c

Created a new `rename_visitor` that has a list of symbol declarations and the references to them. it adds a symbol either as a reference or as a declaration during the visit.
when executing `create_text_edits` it chooses the symbol declaration and references the position belongs to and creates textEdits with the new_text passed by the LSP.
I also updated the `get_ranges` method to get all ranges in a certain line (and not only the first one)

Currently, the renaming of class properties and namespaces is unsupported:
- namespaces need a "customized" edit: `bring ex;` with `externals` as the new word won't become `bring externals;` but `bring ex as externals;`.
- it seems like properties aren't declared/defined as the other symbols- so the current procedure isn't working for them.


- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.