fix(compiler): incorrect lift qualification when accessing a property on a return value of an inflight closure  (#6504)
Fixes #6501

The lift visitor's context now adds information matching the current property with the actual object this property belongs to. This way when qualifying an expression we can verify the expression really belongs to the current property. The result is that something like `(expr_a + expr_b).do()` will qualify only the outer expression with `do()` (if it's a preflight expression). If only the inner expressions are preflight then they will be qualified without any property. This is more correct and doesn't assume any property needs to be qualified when we encounter a preflight expression during the lift visitor.

Specifically this helps us when `some_inflight_closure().do()` the `some_inflight_closure()` call expression isn't a preflight expression so it isn't lifted, but the `some_inflight_closure` part (without the `()` call) is lifted but without the `do()` qualification and defaults to the `handle` qualification which is what we want.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.