chore: ensure wingsdk compiles before running tests (#6514)
Several parts of the SDK reference try to `require`/`require.resolve` `@winglang/sdk/lib/...` while inflight (e.g. for inflight clients). This directory only exists after the SDK has been compiled. This means the SDK tests depend on SDK itself being compiled first. This is unfortunate because we also compile parts of the SDK with tsc as the tests start up, and then vitest itself is also compiling/transpiling.
 
This PR is not doing anything to fix all that, but we at least need to ensure turbo is correctly modeling this dependency to avoid false test failures/successes.

As a side fix, I made eslint allow devDeps in test.config.mts to get rid of red squiggles in my IDE.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.