feat(compiler): support logging stringable values directly (#6521)
Adds support for directly logging stringable values without needing to create a interpolated string:

```
let x = 5;

// before
log("{x}");

// now
log(x);
```

To implement the feature, I had to add a new type to the compiler's type system for representing stringable values. At the moment it more or less behaves like `str | num | bool | ...`. For now it's not really something we have concrete plans to expose in userland (to let users write functions with stringable inputs) but it's something we can consider introducing in the future.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.