fix(compiler): closure transform generated `parent_this` even when not used (#6531)
This is a small code cleanup as part of addressing #4925.

We only generate the `parent_this` variable used in generated closure classes when needed. If the inflight closure never accesses `this` then we don't generated `parent_this`. The code avoids traversing into inner classes inside the closure when searching for `this` accesses, making it a bit cleaner and more efficient.

This relates to #4925 because it solves the problem of accessing `this` (for assignment into `parent_this`) before the `super()` call in a ctor.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.