feat: improve simulator iteration speeds (#6577)
Closes #6540

The main thrust of the change in this PR is to reduce the time spent re-bundling inflight JavaScript code with esbuild. To do that, we take a rough heuristic that most of the time, if none of the files that go into a bundle have changed, then the bundle doesn't need to be regenerated. There are technically some edge cases where this may not be the case, but the heuristic should be accurate enough to handle most local iteration scenarios.

- Update `cloud.Function` simulation to track which files are produced by the esbuild bundle, and consult against the file's timestamps when the resource is choosing its update plan
- Update wingc to avoid re-writing files that haven't changed

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.