chore: avoid full wasm optimizations in development (#6656)
The `-O3` level optimization for wingc's wasm is pretty slow but worth keeping in for the final build.
For local dev it's still valuable to run wasm-opt with some optimization as it greatly reduces the size of the wasm file, which then improves performance when JS loads it. The base level optimization runs in like 500ms on my machine so it's not much of a problem.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.