docs(docs): add link to wing templates (#6684)
Watching the [latest wing update](https://youtu.be/30DVoyAuwP0?si=plqoNGSMhEkNM-PA&t=478) it was great to see the new slack template, but from what I understand the docs don't explain or link to the templates.

I know we can see templates using the CLI, but this is just small change to help users find templates in other ways.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.