feat(docs): test doc examples (#6149)
Closes: https://github.com/winglang/wing/issues/552

## Summary
This PR introduces doc example testing. In its current state you can include the keyword `example` in the Markdown code block identifier similar to how we use the `playground` ability.

If you create a Wing example in the docs and want to have it tested in CI when code changes are made, then all you need to do is something like `"```javascript example" ` and this will mark the code block for testing. 

You can also mark an example for failure if you expect it to not compile by using example meta data like so 
`"```javascript example{valid: false"`

I started adding this to many of our doc examples and found quite a few were broken, so a good chunk of the changes in this PR fix broken examples


- You can also run `pnpm turbo test:doc_examples` from the root of the project to only generate and run doc example tests.

## Further considerations
- How to test examples that require other examples? (I.E. library examples)
- How to test typescript examples

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.