fix(console): prevent default zoom behavior
Prevent the default zoom behavior everywhere in the app.

Since the explorer panel handles the zoom manually, sometimes users may end up zooming the whole app by mistake and end up with the explorer panel covering the whole screen.

This is a big problem since users won't be able to zoom out of it.