docs(docs): added list of platforms to help with discoverability (#6859)
Mentioned by @eladb in https://github.com/winglang/wing/pull/6832

This is an idea to add a new page that lists all the supported platforms by Wing and quick links to dive deeper.
List is split into two, ones that are natively supported by Wing (through CLI), and others that are "external" (requires installation).

Think as we get more community platforms or custom platforms we can add the to this list.

We can also link to this page in the CLI ? (following convo from https://github.com/winglang/wing/pull/6832)


## Checklist

- [x ] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x ] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.